mediofunciona

+++ srcs/requirements/wordpress/script.sh
@@ -1,21 +1,27 @@
-#!/bin/bash
-WP_ADMIN_E=$(cut -d '=' -f2 /run/secrets/wp_admin_e.txt)
-WP_ADMIN_N=$(cut -d '=' -f2 /run/secrets/wp_admin_n.txt)
-WP_ADMIN_P=$(cut -d '=' -f2 /run/secrets/wp_admin_p.txt)
-WP_USER_P=$(cut -d '=' -f2 /run/secrets/wp_user_p.txt)
-DB_PWD=$(cut -d '=' -f2 /run/secrets/db_pwd.txt)
+#!/bin/sh
 
-sleep 3
+username=$WP_USERNAME
+userpwd=$(cut -f2 /run/secrets/wp-user-password)
+adminuser=$(cut -f2 /run/secrets/wp-admin-user)
+adminpwd=$(cut -f2 /run/secrets/wp-admin-password)
+dbname=$(cut -f2 /run/secrets/db-user)
+dbpwd=$(cut -f2 /run/secrets/db-password)
+
 if ! [ -e /var/www/html/wp-config.php ]; then
-    echo DOWLOADING WORDPRESS
     cd /var/www/html
-    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-    chmod +x wp-cli.phar
-    ./wp-cli.phar core download --allow-root
-    ./wp-cli.phar config create --dbname="$DB_NAME" --dbuser="$DB_USER" --dbpass="$DB_PWD" --dbhost=mariadb --allow-root
-    ./wp-cli.phar core install --url="$DOMAIN_N" --title="$WP_TITLE" --admin_user="$WP_ADMIN_N" --admin_password="$WP_ADMIN_P" --admin_email="$WP_ADMIN_E" --allow-root
-    ./wp-cli.phar user create  "$WP_USER_N" "$WP_USER_E" --user_pass="$WP_USER_P" --role="$WP_USER_R" --allow-root
+    rm -rf /var/www/html/*
+    if ! [ -e /usr/local/bin/wp ]; then
+        curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+        chmod +x wp-cli.phar
+        mv wp-cli.phar /usr/local/bin/wp
+    fi
+    wp core download --allow-root
+    wp config create --dbname=wordpress --dbuser=$dbname --dbpass=$dbpwd --dbhost=mariadb --allow-root
+    wp core install --url=$DOMAIN_NAME --title=inception --admin_user=$adminuser --admin_password=$adminpwd --admin_email=$adminuser@example.com --allow-root
+    wp user create $username $username@example.com --role=author --user_pass=$userpwd --allow-root
+    wp theme install twentysixteen --activate --allow-root
 else
-    echo WORDPRESS ALREADY INSTALLED
+    echo "Wordpress already installed"
 fi
-php-fpm83 -F
+
+exec "php-fpm8.4" "-F"
